### 2017-12-17 #513 @whedon
[REVIEW]: cleanEHR: The critical care EHR analysis tools
**Submitting author:** @sinanshi (<a href="http://orcid.org/0000-0001-6935-0429">Sinan Shi</a>)
**Repository:** <a href="https://github.com/ropensci/cleanEHR" target ="_blank">https://github.com/ropensci/cleanEHR</a>
**Version:** v1.0
