### 2017-12-17 #513 @arfon
Accepting because of submission to rOpenSci: https://github.com/ropensci/onboarding/issues/102
