### 2017-12-17 #513 @arfon
@sinanshi - I've fixed up some issues with your `paper.md` here https://github.com/ropensci/cleanEHR/pull/151 .

Also, since you submitted originally, our [submission guidelines](http://joss.theoj.org/about#paper_structure) have been updated a fair bit. Please take a look at the section 'What should my paper contain?' here http://joss.theoj.org/about#paper_structure

Finally, it looks like you're missing a bibliography. Please make sure you cite related work using the standard RMarkdown way: can read how to do that [here](http://rmarkdown.rstudio.com/authoring_bibliographies_and_citations.html))
