### 2017-12-17 #513 @sinanshi
Thanks @arfon, could you have a look if it is all OK now.
