### 2017-12-18 #513 @arfon
Thanks @sinanshi  - I still think this paper could do with a little more detail. From the text of your paper it's still not very clear to me what this software does! Please consider adding just a little more prose, perhaps with an example use case to address this:

> - A summary describing the high-level functionality and purpose of the software for a diverse, non-specialist audience

Also, if there are any research projects using this software then please describe them too:

> - Mentions (if applicable) of any ongoing research projects using the software or recent scholarly publications enabled by it
