### 2017-12-19 #513 @sinanshi
Thank you @arfon for your suggestions. I have updated the paper, adding the high-level functionality description. Regarding other related research project, there is no published research result so far using this package, despite many at the early stage of their research.
